requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pyperclip>=1.8.2
pathlib
flask>=2.3.0
werkzeug>=2.3.0 